+++ python/cuxfilter/charts/core/core_widget.py
@@ -1,4 +1,9 @@
 from typing import Dict
+import logging
+from panel.config import panel_extension
+import panel as pn
+
+from ...layouts import chart_view
 
 
 class BaseWidget:
@@ -18,6 +23,7 @@ class BaseWidget:
     max_value: float = 0.0
     label_map: Dict[str, str] = None
     use_data_tiles = False
+    _initialized = False
 
     @property
     def name(self):
@@ -74,8 +80,30 @@ class BaseWidget:
             self.label_map = {v: k for k, v in self.label_map.items()}
             params.pop("label_map")
 
+    def _repr_mimebundle_(self, include=None, exclude=None):
+        view = self.view()
+        if self._initialized and panel_extension._loaded:
+            return view._repr_mimebundle_(include, exclude)
+
+        if self._initialized is False:
+            logging.warning(
+                "dashboard has not been initialized."
+                "Please run cuxfilter.dashboard.Dashboard([...charts])"
+                " to view this object in notebook"
+            )
+
+        if panel_extension._loaded is False:
+            logging.warning(
+                "notebooks assets not loaded."
+                "Please run cuxfilter.load_notebooks_assets()"
+                " to view this object in notebook"
+            )
+            if isinstance(view, pn.Column):
+                return view.pprint()
+        return None
+
     def view(self):
-        return self.chart
+        return chart_view(self.chart, width=self.width)
 
     def add_event(self, event, callback):
         self.chart.on_event(event, callback)